fix: encode should_reexecute as i64 to avoid colliding with bci marker
deopt_args pushed should_reexecute as an i32, immediately before the
duplicated-bci pair (also i32, used as a search key by the LLVM side to
locate the current scope). should_reexecute is only ever 0 or 1, which
can equal a small bci value, so a same-width should_reexecute risks
being mistaken for one half of the duplicated-bci pair.

Push should_reexecute as i64 instead so it can't be confused with the
i32 bci marker, and update parse_stackmap's read side to match using
the existing wide-constant accessor (StackMapUtil::getConstantUlong).

Companion fix in jeandle-llvm's JeandleNarrowOopMarker (synthetic deopt
bundle for call sites with no prior deopt info) needs the same width
change; carried by CO_BRANCH below until it lands in jeandle-llvm main.

CO_REPO: https://github.com/xugeeeee/jeandle-llvm.git
CO_BRANCH: fix/narrow-oop-marker-missing-reexecute-flag

+++ src/hotspot/share/jeandle/jeandleAbstractInterpreter.cpp
@@ -269,7 +269,11 @@ llvm::SmallVector<llvm::Value*> JeandleVMState::deopt_args(llvm::IRBuilder<>& bu
   // should_reexecute is explicitly set by intrinsic lowering (e.g. addExact's
   // overflow trap) to force reexecution of the current bci on deopt, matching
   // C2's should_reexecute semantics for intrinsic-emitted uncommon traps.
-  args.push_back(builder.getInt32(should_reexecute ? 1 : 0));
+  //
+  // Pushed as i64 (not i32) so it can't be confused with the duplicated-BCI
+  // marker below: the marker is identified by two adjacent i32 values, and
+  // should_reexecute (0 or 1) can easily collide with a small bci value.
+  args.push_back(builder.getInt64(should_reexecute ? 1 : 0));
 
   // Duplicate the BCI as a BCI marker for the LLVM backend.
   // Keep TestScopeValues.java in sync with this duplicated-BCI convention.
